+++ test/specs/spray.spec.js
@@ -1,8 +1,8 @@
 /* globals it, describe, before, beforeEach, expect, chai, sinonChai, sinon, fixture */
 /* eslint no-unused-expressions: 0 */
 
-import fire from '../../src/fire';
-import { load, spray } from '../../src/spray';
+import { fire, load } from '../../src/fire';
+import spray from '../../src/spray';
 import keyCode from '../../src/events/keyboard.utils';
 import { center } from '../../src/utils/position.utils';
 
@@ -21,9 +21,11 @@ describe('load and spray', () => {
     document.body.style.margin = '20px';
     element.style.width = '100px';
     element.style.height = '100px';
+    sinon.spy(console, 'warn');
   });
 
   afterEach(() => {
+    console.warn.restore();
     sandbox.restore();
   });
 
@@ -35,23 +37,24 @@ describe('load and spray', () => {
 
     it('should be a higher order function', () => {
       expect(typeof load).to.eql('function');
-      expect(typeof load('MouseEvent', 'click')).to.eql('function');
+      expect(typeof load('click')).to.eql('function');
     });
 
     it('should return a Promise Object', () => {
-      expect(typeof load('MouseEvent', 'click')({})).to.eql('object');
+      expect(typeof load('click')({})).to.eql('object');
     });
 
     it('should return a new event from promise', (done) => {
-      load('MouseEvent', 'click')({}).then(({event}) => {
+      load('click')({}).then(({event}) => {
         expect(event.clientX).to.eql(0);
         done();
       });
     });
 
     it('should return a new event from promise from current element', (done) => {
-      load('MouseEvent', 'click', element)({}).then(({event}) => {
+      load('click', element)({}).then(({event}) => {
         expect(event.clientX).to.eql(20);
+        expect(console.warn).not.to.have.been.called;
         done();
       });
     });
@@ -60,6 +63,16 @@ describe('load and spray', () => {
       load('MouseEvent', 'click', document, center())({})
       .then(({event}) => {
         expect(event.clientX).to.not.eql(0);
+        expect(console.warn).to.have.been.called;
+        done();
+      });
+    });
+
+    it('should return a new event from that promise callback', (done) => {
+      load('click', document, center())({})
+      .then(({event}) => {
+        expect(event.clientX).to.not.eql(0);
+        expect(console.warn).not.to.have.been.called;
         done();
       });
     });
@@ -73,7 +86,7 @@ describe('load and spray', () => {
     it('should take a load hoc and fire 10 times as default', (done) => {
       const eventSpy = sinon.spy();
       document.addEventListener('click', eventSpy);
-      const clip = load('MouseEvent', 'click', document, center());
+      const clip = load('click', document, center());
       spray(clip).then(() => {
         const times = eventSpy.printf('%c');
         expect(eventSpy.called).to.be.true;
@@ -85,7 +98,7 @@ describe('load and spray', () => {
     it('should take a load hoc and fire 15 times', (done) => {
       const eventSpy = sinon.spy();
       document.addEventListener('click', eventSpy);
-      const clip = load('MouseEvent', 'click', document, center());
+      const clip = load('click', document, center());
       spray(clip, {
         steps: 15
       }).then(() => {
@@ -99,7 +112,7 @@ describe('load and spray', () => {
     it('should take a load hoc and fire 4 times', (done) => {
       const eventSpy = sinon.spy();
       document.addEventListener('click', eventSpy);
-      const clip = load('MouseEvent', 'click', document, center());
+      const clip = load('click', document, center());
       spray(clip, {
         steps: 4
       }).then(() => {
@@ -114,7 +127,7 @@ describe('load and spray', () => {
       const eventSpy = sinon.spy();
       const pathSpy = sinon.spy();
       document.addEventListener('click', eventSpy);
-      const clip = load('MouseEvent', 'click', document, center());
+      const clip = load('click', document, center());
       spray(clip, {
         steps: 4,
         path: pathSpy,
@@ -132,7 +145,7 @@ describe('load and spray', () => {
     describe('should update clientX and clientY path', () => {
       it('when it it a function', (done) => {
         const { clientX: centerX, clientY: centerY } = center();
-        const clip = load('MouseEvent', 'click', document, center());
+        const clip = load('click', document, center());
         spray(clip, {
           steps: 4,
           path: ({clientX, clientY}) => {
@@ -150,7 +163,7 @@ describe('load and spray', () => {
 
       it('when it it an object', (done) => {
         const { clientX: centerX, clientY: centerY } = center();
-        const clip = load('MouseEvent', 'click', document, center());
+        const clip = load('click', document, center());
         spray(clip, {
           steps: 4,
           path: {clientX: 5, clientY: 2},
@@ -166,7 +179,7 @@ describe('load and spray', () => {
       const eventSpy = sinon.spy();
       const tickSpy = sinon.spy();
       document.addEventListener('click', eventSpy);
-      const clip = load('MouseEvent', 'click', document, center());
+      const clip = load('click', document, center());
       spray(clip, {
         steps: 4,
         tick: tickSpy,
@@ -188,7 +201,7 @@ describe('load and spray', () => {
     describe('KeyboardEvent', () => {
       it('should fire multiple keys in a sequence from path function', (done) => {
         const keysPressed = [];
-        const clip = load('KeyboardEvent', 'keypress', document, keyCode('a'));
+        const clip = load('keypress', document, keyCode('a'));
         spray(clip, {
           steps: 4,
           path: ({key}) => {
@@ -210,7 +223,7 @@ describe('load and spray', () => {
         const convertKeys = (...keys) => keys.map(key => keyCode(key));
         const convertedKeys = convertKeys('a', 'b', 'c', 'd');
 
-        const clip = load('KeyboardEvent', 'keypress', document, keyCode('a'));
+        const clip = load('keypress', document, keyCode('a'));
         spray(clip, {
           steps: 4,
           path: convertedKeys,
@@ -235,15 +248,15 @@ describe('load and spray', () => {
 
       it('a move event', (done) => {
         const { clientX: centerX, clientY: centerY } = center();
-        fire('MouseEvent', 'moveenter', document, center());
-        const clip = load('MouseEvent', 'mousemove', document, center());
+        fire('mouseenter', document, center());
+        const clip = load('mousemove', document, center());
         spray(clip, {
           steps: 4,
           path: {clientX: 5, clientY: 2},
         }).then(({clientX, clientY}) => {
           expect(clientX).to.eql(Math.floor(centerX + (3 * 5)));
           expect(clientY).to.eql(Math.floor(centerY + (3 * 2)));
-          fire('MouseEvent', 'moveleave', document, {clientX, clientY});
+          fire('moveleave', document, {clientX, clientY});
           done();
         });
       });